- Added husky and check .gitignore existence before deleting it

+++ cli.js
@@ -40,7 +40,10 @@ const editFiles = async projectName => {
   file.set('name', projectName);
   file.save();
 
-  fs.unlinkSync(`${__dirname}/starter-kit/.gitignore`);
+  if(fs.existsSync(`${__dirname}/starter-kit/.gitignore`)) {
+    fs.unlinkSync(`${__dirname}/starter-kit/.gitignore`);
+  }
+
   fs.copyFile(
     `${__dirname}/starter-kit/gitignore`,
     `${__dirname}/starter-kit/.gitignore`,
